+++ model_ed4.py
@@ -88,7 +88,7 @@ imgs_all = []
 for idx, r in train_only_all_points.iterrows():
     imgs_all.append(
         np.array(r["Image"].split())
-        .astype(np.int64)
+        .astype(np.uint8)
         .reshape(IMAGE_WIDTH, IMAGE_HEIGHT, 1)
     )
 imgs_all = np.array(imgs_all)
@@ -99,7 +99,7 @@ imgs_mounth_points_all = []
 for idx, r in mouth_left_corner_points.iterrows():
     imgs_mounth_points_all.append(
         np.array(r["Image"].split())
-        .astype(np.int64)
+        .astype(np.uint8)
         .reshape(IMAGE_WIDTH, IMAGE_HEIGHT, 1)
     )
 imgs_mount_points_all = np.array(imgs_mounth_points_all)
